Move chat-input.test.tsx to components/chat directory
